chore(skeleton-image): migrate skeleton image to new folder
