+++ packages/vant/src/skeleton/index.less
@@ -3,8 +3,6 @@
   --van-skeleton-paragraph-background: var(--van-active-color);
   --van-skeleton-paragraph-margin-top: var(--van-padding-sm);
   --van-skeleton-title-width: 40%;
-  --van-skeleton-image-size: 96px;
-  --van-skeleton-image-radius: 24px;
   --van-skeleton-duration: 1.2s;
 }
 
@@ -44,26 +42,6 @@
   &--animate {
     animation: van-skeleton-blink var(--van-skeleton-duration) ease-in-out
       infinite;
-  }
-
-  &-image {
-    display: flex;
-    width: var(--van-skeleton-image-size);
-    height: var(--van-skeleton-image-size);
-    align-items: center;
-    justify-content: center;
-    background: var(--van-active-color);
-
-    &--round {
-      border-radius: var(--van-skeleton-image-radius);
-    }
-
-    &__icon {
-      width: calc(var(--van-skeleton-image-size) / 2);
-      height: calc(var(--van-skeleton-image-size) / 2);
-      font-size: calc(var(--van-skeleton-image-size) / 2);
-      color: var(--van-gray-5);
-    }
   }
 }
