+++ packages/vant/src/skeleton/index.ts
@@ -1,11 +1,9 @@
 import _Skeleton from './Skeleton';
-import _SkeletonImage from './SkeletonImage';
 import _SkeletonTitle from './SkeletonTitle';
 import _SkeletonParagraph from './SkeletonParagraph';
 
 import { withInstall } from '../utils';
 
-export const SkeletonImage = withInstall(_SkeletonImage);
 export const SkeletonTitle = withInstall(_SkeletonTitle);
 export const SkeletonParagraph = withInstall(_SkeletonParagraph);
 export const Skeleton = withInstall(_Skeleton);
@@ -16,10 +14,6 @@ export default Skeleton;
 export { skeletonProps } from './Skeleton';
 export type { SkeletonProps } from './Skeleton';
 
-// SkeletonImage
-export { skeletonImageProps } from './SkeletonImage';
-export type { SkeletonImageProps } from './SkeletonImage';
-
 // SkeletonParagraph
 export { skeletonParagraphProps } from './SkeletonParagraph';
 export type { SkeletonParagraphProps } from './SkeletonParagraph';
@@ -28,16 +22,11 @@ export type { SkeletonParagraphProps } from './SkeletonParagraph';
 export { skeletonTitleProps } from './SkeletonTitle';
 export type { SkeletonTitleProps } from './SkeletonTitle';
 
-export type {
-  SkeletonThemeVars,
-  SkeletonImageShape,
-  SkeletonAvatarShape,
-} from './types';
+export type { SkeletonThemeVars } from './types';
 
 declare module 'vue' {
   export interface GlobalComponents {
     VanSkeleton: typeof Skeleton;
-    VanSkeletonImage: typeof SkeletonImage;
     VanSkeletonTitle: typeof SkeletonTitle;
     VanSkeletonParagraph: typeof SkeletonParagraph;
   }
